Consolidate browser abstraction code
This includes functions to manage remote browser (previously at
`chromefleet.py`) as well page and element wrappers (previously at
`zen_distill.py`).
