+++ getgather/mcp/amazonca.py
@@ -7,6 +7,7 @@ from typing import Any, cast
 import zendriver as zd
 from loguru import logger
 
+from getgather.browser import get_url, page_query_selector
 from getgather.mcp.dpage import (
     remote_zen_dpage_mcp_tool,
     remote_zen_dpage_with_action,
@@ -14,9 +15,7 @@ from getgather.mcp.dpage import (
 from getgather.mcp.registry import GatherMCP
 from getgather.zen_distill import (
     convert,
-    get_url,
     load_distillation_patterns,
-    page_query_selector,
     run_distillation_loop,
 )
