+++ getgather/mcp/doordash.py
@@ -4,13 +4,13 @@ from typing import Any, cast
 import zendriver as zd
 from loguru import logger
 
+from getgather.browser import zen_navigate_with_retry
 from getgather.mcp.dpage import (
     remote_zen_dpage_mcp_tool,
     remote_zen_dpage_with_action,
 )
 from getgather.mcp.registry import GatherMCP
 from getgather.mcp.utils import retry_with_navigation
-from getgather.zen_distill import zen_navigate_with_retry
 
 doordash_mcp = GatherMCP(brand_id="doordash", name="Doordash MCP")
