+++ getgather/mcp/tokopedia.py
@@ -6,6 +6,7 @@ from urllib.parse import quote, urlparse
 import zendriver as zd
 from loguru import logger
 
+from getgather.browser import page_query_selector, zen_navigate_with_retry
 from getgather.mcp.dpage import (
     remote_zen_dpage_mcp_tool,
     remote_zen_dpage_with_action,
@@ -13,7 +14,6 @@ from getgather.mcp.dpage import (
 from getgather.mcp.registry import GatherMCP
 from getgather.mcp.utils import retry_with_navigation
 from getgather.zen_actions import parse_response_json
-from getgather.zen_distill import page_query_selector, zen_navigate_with_retry
 
 tokopedia_mcp = GatherMCP(brand_id="tokopedia", name="Tokopedia MCP")
