+++ getgather/zen_distill.py
@@ -1,7 +1,6 @@
 import asyncio
 import json
 import os
-import random
 import re
 import urllib.parse
 from dataclasses import dataclass
@@ -12,15 +11,22 @@ from typing import Any, cast
 from urllib.parse import urlunparse
 
 import sentry_sdk
-import websockets
 import zendriver as zd
 from bs4 import BeautifulSoup
 from bs4.element import Tag
 from loguru import logger
 from nanoid import generate
-from zendriver.core.connection import ProtocolException
 
-from getgather.browser.chromefleet import create_remote_browser, terminate_remote_browser
+from getgather.browser import (
+    create_remote_browser,
+    get_new_page,
+    page_batch_extract,
+    page_query_selector,
+    safe_close_page,
+    terminate_remote_browser,
+    wait_for_ready_state,
+    zen_navigate_with_retry,
+)
 from getgather.config import FRIENDLY_CHARS, settings
 
 
@@ -37,35 +43,7 @@ class Match:
     distilled: str
 
 
-@dataclass
-class ElementConfig:
-    """Configuration for element typing operations."""
-
-    # Typing configuration
-    typing_clear_delay: float = 0.1
-    typing_char_delay_min: float = 0.01
-    typing_char_delay_max: float = 0.05
-
-
 ConversionResult = list[dict[str, str | list[str]]]
-
-_CREDENTIALS_BLOCK_SCRIPT = r"""
-(() => {
-  "use strict";
-
-  if (!window.CredentialsContainer || !navigator.credentials) return;
-
-  Object.defineProperty(CredentialsContainer.prototype, "get", {
-    configurable: true,
-    writable: true,
-    value: function () {
-      return Promise.reject(
-        new DOMException("Credentials API blocked", "NotAllowedError")
-      );
-    },
-  });
-})();
-"""
 
 NETWORK_ERROR_PATTERNS = (
     "err-timed-out",
@@ -226,32 +204,6 @@ def _safe_fragment(value: str) -> str:
     return fragment or "distill"
 
 
-async def wait_for_ready_state(
-    page: zd.Tab,
-    timeout: int = 10,
-) -> bool:
-    """
-    Waits for the page to reach a certain ready state (interactive or complete).
-    :param timeout: The maximum number of seconds to wait.
-    :type timeout: int
-    :raises asyncio.TimeoutError: If the timeout is reached before the ready state is reached.
-    :return: True if the ready state is reached.
-    :rtype: bool
-    """
-    loop = asyncio.get_event_loop()
-    start_time = loop.time()
-
-    while True:
-        state = await page.evaluate("document.readyState")
-        if state == "interactive" or state == "complete":
-            return True
-
-        if loop.time() - start_time > timeout:
-            raise asyncio.TimeoutError("time ran out while waiting for load page until %s" % state)
-
-        await asyncio.sleep(0.1)
-
-
 async def capture_page_artifacts(
     page: zd.Tab,  # type: ignore[name-defined]
     *,
@@ -347,539 +299,6 @@ async def zen_report_distill_error(
                 )
 
             sentry_sdk.capture_exception(error)
-
-
-async def zen_navigate_with_retry(page: zd.Tab, url: str, wait_for_ready: bool = True) -> zd.Tab:
-    """Navigate to URL with retry logic for resilient navigation.
-
-    Args:
-        page: Zendriver tab to navigate
-        url: URL to navigate to
-        wait_for_ready: Whether to wait for page ready state (default True).
-            Set to False for simple pages that load instantly.
-
-    Returns:
-        The page after successful navigation
-
-    Raises:
-        Exception: If navigation fails after all retries
-    """
-    MAX_RETRIES = 3
-    FIRST_TIMEOUT = 45  # seconds, extended for first attempt
-    NORMAL_TIMEOUT = 30  # seconds, for retry attempts
-
-    last_error: Exception | None = None
-    for attempt in range(MAX_RETRIES):
-        timeout = FIRST_TIMEOUT if attempt == 0 else NORMAL_TIMEOUT
-        try:
-
-            async def navigate_and_wait() -> zd.Tab:
-                _frame_id, _loader_id, error_text = await page.send(zd.cdp.page.navigate(url))
-
-                # Check for navigation errors (connection refused, DNS failure, SSL errors, etc.)
-                if error_text:
-                    raise ConnectionError(f"Navigation failed: {error_text}")
-
-                if not wait_for_ready:
-                    return page
-
-                # Wait for page to be interactive
-                try:
-                    await wait_for_ready_state(page)
-                except Exception:
-                    # If wait fails, that's okay - page might already be loaded
-                    pass
-                return page
-
-            result = await asyncio.wait_for(navigate_and_wait(), timeout=timeout)
-            return result
-        except Exception as error:
-            last_error = error
-            if attempt < MAX_RETRIES - 1:
-                logger.warning(
-                    f"Navigation to {url} failed (attempt {attempt + 1}/{MAX_RETRIES}): {error}. "
-                    f"Retrying in 1 second..."
-                )
-                await asyncio.sleep(1)
-            else:
-                logger.error(f"Failed to navigate to {url} after {MAX_RETRIES} attempts")
-
-    # This should never be reached, but satisfies type checker
-    raise last_error or Exception(f"Failed to navigate to {url}")
-
-
-async def get_new_page(browser: zd.Browser) -> zd.Tab:
-    page = await browser.get("about:blank", new_tab=True)
-
-    async def handle_request(event: zd.cdp.fetch.RequestPaused) -> None:
-        resource_type = event.resource_type
-        request_url = event.request.url
-
-        deny = resource_type in [
-            zd.cdp.network.ResourceType.IMAGE,
-            zd.cdp.network.ResourceType.MEDIA,
-            zd.cdp.network.ResourceType.FONT,
-        ]
-
-        if not deny:
-            try:
-                await page.send(zd.cdp.fetch.continue_request(request_id=event.request_id))
-            except (ProtocolException, websockets.ConnectionClosedError) as e:
-                if isinstance(e, ProtocolException) and (
-                    "Invalid state for continueInterceptedRequest" in str(e)
-                    or "Invalid InterceptionId" in str(e)
-                    or "Fetch domain is not enabled" in str(e)
-                ):
-                    logger.debug(
-                        f"Request already processed or invalid interception ID: {request_url}"
-                    )
-                elif isinstance(e, websockets.ConnectionClosedError):
-                    logger.debug(f"Page closed while continuing request: {request_url}")
-                else:
-                    raise
-            return
-
-        logger.trace(f" DENY resource: {request_url}")
-
-        try:
-            await page.send(
-                zd.cdp.fetch.fail_request(
-                    request_id=event.request_id,
-                    error_reason=zd.cdp.network.ErrorReason.BLOCKED_BY_CLIENT,
-                )
-            )
-        except (ProtocolException, websockets.ConnectionClosedError) as e:
-            if isinstance(e, ProtocolException) and (
-                "Invalid state for continueInterceptedRequest" in str(e)
-                or "Invalid InterceptionId" in str(e)
-                or "Fetch domain is not enabled" in str(e)
-            ):
-                logger.debug(f"Request already processed or invalid interception ID: {request_url}")
-            elif isinstance(e, websockets.ConnectionClosedError):
-                logger.debug(f"Page closed while blocking request: {request_url}")
-            else:
-                raise
-
-    # Enable fetch domain to intercept requests. Will be overridden if proxy auth is set up.
-    await page.send(zd.cdp.fetch.enable())
-    page.add_handler(zd.cdp.fetch.RequestPaused, handle_request)  # type: ignore[reportUnknownMemberType]
-
-    # Block the entire Credentials API before page for passkeys
-    # Page domain must be enabled before addScriptToEvaluateOnNewDocument, otherwise CDP will reject the script
-    await page.send(zd.cdp.page.enable())
-    await page.send(
-        zd.cdp.page.add_script_to_evaluate_on_new_document(
-            source=_CREDENTIALS_BLOCK_SCRIPT, run_immediately=True
-        )
-    )
-
-    return page
-
-
-async def safe_close_page(page: zd.Tab) -> None:
-    """Safely close a page by disabling fetch domain first to prevent orphaned tasks.
-
-    When page.close() is called while fetch handlers are pending, it can leave
-    orphaned tasks waiting for CDP responses that will never arrive. This function
-    disables the fetch domain first to clean up handlers before closing.
-    """
-    try:
-        # Disable fetch domain to cancel pending request handlers
-        await page.send(zd.cdp.fetch.disable())
-        logger.debug("Fetch domain disabled before page close")
-    except (ProtocolException, websockets.ConnectionClosedError) as e:
-        # Page/connection already closed, which is fine
-        logger.debug(f"Could not disable fetch (connection already closed): {e}")
-    except Exception as e:
-        # Log but don't fail - we still want to close the page
-        logger.warning(f"Unexpected error disabling fetch domain: {e}")
-
-    try:
-        target = getattr(page, "target", None)
-        raw_target_id = getattr(target, "target_id", None) or getattr(page, "target_id", None)
-        normalized_target_id = (
-            zd.cdp.target.TargetID(raw_target_id.split("@", 1)[-1])
-            if isinstance(raw_target_id, str)
-            else None
-        )
-
-        if normalized_target_id:
-            await page.send(zd.cdp.target.close_target(target_id=normalized_target_id))
-        else:
-            await page.close()
-        logger.warning("Page closed successfully")
-    except Exception as e:
-        logger.warning(f"Error closing page: {e}")
-
-
-class Element:
-    """Wrapper to handle both CSS and XPath selector differences for browser elements."""
-
-    def __init__(
-        self,
-        element: zd.Element,
-        css_selector: str | None = None,
-        xpath_selector: str | None = None,
-        config: ElementConfig | None = None,
-    ):
-        self.element = element
-        self.tag = element.tag
-        self.page = element.tab
-        self.css_selector = css_selector
-        self.xpath_selector = xpath_selector
-        self.config = config or ElementConfig()
-
-    async def inner_html(self) -> str:
-        return await self.element.get_html()
-
-    async def inner_text(self) -> str:
-        return self.element.text_all
-
-    async def is_visible(self) -> bool:
-        try:
-            selector = self.xpath_selector or self.css_selector
-            if not selector:
-                return False
-
-            # Optimized: Single JavaScript evaluation with minimal operations
-            escaped_selector = selector.replace("\\", "\\\\").replace('"', '\\"')
-
-            if self.xpath_selector:
-                # Use a more efficient check - combine all checks in one expression
-                js_code = f"""
-                    (() => {{
-                        const el = document.evaluate("{escaped_selector}", document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null).singleNodeValue;
-                        if (!el) return false;
-                        const s = window.getComputedStyle(el);
-                        return s.visibility !== "hidden" && s.display !== "none" && el.getBoundingClientRect().width > 0 && el.getBoundingClientRect().height > 0;
-                    }})()
-                """
-            else:
-                # Optimized: Single expression, no intermediate variables
-                js_code = f"""
-                    (() => {{
-                        const el = document.querySelector("{escaped_selector}");
-                        if (!el) return false;
-                        const s = window.getComputedStyle(el);
-                        return s.visibility !== "hidden" && s.display !== "none" && el.getBoundingClientRect().width > 0 && el.getBoundingClientRect().height > 0;
-                    }})()
-                """
-
-            return bool(await self.page.evaluate(js_code))
-        except Exception as e:
-            logger.debug(f"is_visible check failed: {e}")
-            return False
-
-    async def click(self) -> None:
-        if self.css_selector:
-            await self.css_click()
-        else:
-            await self.xpath_click()
-        await asyncio.sleep(0.25)
-
-    async def select_option(self, value: str) -> None:
-        # Only support CSS selectors for now
-        if not self.css_selector:
-            logger.warning("Cannot perform CSS select_option: no css_selector available")
-            return
-        logger.debug(f"Attempting JavaScript CSS select_option for {self.css_selector}")
-        try:
-            escaped_selector = self.css_selector.replace("\\", "\\\\").replace('"', '\\"')
-            value_selector = f"option[value='{value}']"
-            js_code = f"""
-                (() => {{
-                    const select = document.querySelector("{escaped_selector}");
-                    const option = select?.querySelector("{value_selector}");
-                    if (!select || !option) return false;
-
-                    // Scroll into view
-                    select.scrollIntoView({{ block: "center" }});
-
-                    // Open dropdown (if needed)
-                    select.dispatchEvent(new PointerEvent("pointerdown", {{ bubbles: true }}));
-                    select.dispatchEvent(new PointerEvent("pointerup", {{ bubbles: true }}));
-                    select.dispatchEvent(new MouseEvent("click", {{ bubbles: true, cancelable: true, view: window }}));
-
-                    // Select the option
-                    option.selected = true;
-
-                    // Trigger change event
-                    select.dispatchEvent(new Event("change", {{ bubbles: true }}));
-
-                    return true;
-                }})();
-            """
-            result = await self.page.evaluate(js_code)
-            if result:
-                logger.info(f"JavaScript CSS select_option succeeded for {self.css_selector}")
-                return
-            else:
-                logger.warning(
-                    f"JavaScript CSS select_option could not find element {self.css_selector}"
-                )
-        except Exception as js_error:
-            logger.error(f"JavaScript CSS select_option failed: {js_error}")
-
-        await asyncio.sleep(0.25)
-
-    async def check(self) -> None:
-        logger.error("TODO: Element#check")
-        await asyncio.sleep(0.25)
-
-    async def type_text(self, text: str) -> None:
-        await self.element.clear_input_by_deleting()
-        await asyncio.sleep(self.config.typing_clear_delay)
-        await self.element.clear_input()
-        await asyncio.sleep(self.config.typing_clear_delay)
-        for char in text:
-            await self.element.send_keys(char)
-            await asyncio.sleep(
-                random.uniform(self.config.typing_char_delay_min, self.config.typing_char_delay_max)
-            )
-
-    async def css_click(self) -> None:
-        if not self.css_selector:
-            logger.warning("Cannot perform CSS click: no css_selector available")
-            return
-        logger.debug(f"Attempting JavaScript CSS click for {self.css_selector}")
-        try:
-            escaped_selector = self.css_selector.replace("\\", "\\\\").replace('"', '\\"')
-            js_code = f"""
-            (() => {{
-                const selector = "{escaped_selector}";
-                function findInDocument(doc) {{
-                    try {{
-                        const el = doc.querySelector(selector);
-                        if (el) return el;
-                    }} catch (e) {{
-                        // Cross-origin iframe → skip
-                    }}
-                    // Look inside all iframes of this document
-                    const iframes = doc.querySelectorAll("iframe");
-                    for (const frame of iframes) {{
-                        try {{
-                            const childDoc = frame.contentDocument || frame.contentWindow.document;
-                            const found = findInDocument(childDoc);   // recursion
-                            if (found) return found;
-                        }} catch (e) {{
-                            // Cross-origin iframe → skip
-                        }}
-                    }}
-                    return null;
-                }}
-                const element = findInDocument(document);
-                if (!element) return false;
-                element.scrollIntoView({{ block: "center" }});
-                element.dispatchEvent(new PointerEvent("pointerdown", {{ bubbles: true }}));
-                element.dispatchEvent(new PointerEvent("pointerup", {{ bubbles: true }}));
-                element.dispatchEvent(new MouseEvent("click", {{ bubbles: true, cancelable: true, view: window }}));
-                return true;
-            }})()
-            """
-            result = await self.page.evaluate(js_code)
-            if result:
-                logger.info(f"JavaScript CSS click succeeded for {self.css_selector}")
-                return
-            else:
-                logger.warning(f"JavaScript CSS click could not find element {self.css_selector}")
-        except Exception as js_error:
-            logger.error(f"JavaScript CSS click failed: {js_error}")
-
-    async def xpath_click(self) -> None:
-        if not self.xpath_selector:
-            logger.warning(f"Cannot perform XPath click: no xpath_selector available")
-            return
-        logger.debug(f"Attempting JavaScript XPath click for {self.xpath_selector}")
-        try:
-            escaped_selector = self.xpath_selector.replace("\\", "\\\\").replace('"', '\\"')
-            js_code = f"""
-            (() => {{
-                let element = document.evaluate("{escaped_selector}", document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null).singleNodeValue;
-                if (!element) return false;
-                element.scrollIntoView({{ block: "center" }});
-                element.dispatchEvent(new PointerEvent("pointerdown", {{ bubbles: true }}));
-                element.dispatchEvent(new PointerEvent("pointerup", {{ bubbles: true }}));
-                element.dispatchEvent(new MouseEvent("click", {{ bubbles: true, cancelable: true, view: window }}));
-                return true;
-            }})()
-            """
-            result = await self.page.evaluate(js_code)
-            if result:
-                logger.info(f"JavaScript XPath click succeeded for {self.xpath_selector}")
-                return
-            else:
-                logger.warning(
-                    f"JavaScript XPath click could not find element {self.xpath_selector}"
-                )
-        except Exception as js_error:
-            logger.error(f"JavaScript XPath click failed: {js_error}")
-
-
-async def page_query_selector(
-    page: zd.Tab,
-    selector: str,
-    timeout: float = 0,
-    iframe_selector: str | None = None,
-    skip_visibility_check: bool = False,
-    config: ElementConfig | None = None,
-) -> Element | None:
-    try:
-        if selector.startswith("//"):
-            elements = await page.xpath(selector, timeout)
-            if elements and len(elements) > 0:
-                element = Element(elements[0], xpath_selector=selector, config=config)
-                if skip_visibility_check or await element.is_visible():
-                    return element
-            return None
-
-        if iframe_selector is not None:
-            element = await page.select_all(selector, timeout=timeout, include_frames=True)
-            if element and len(element) > 0:
-                element = Element(element[0], css_selector=selector, config=config)
-                if skip_visibility_check or await element.is_visible():
-                    return element
-        else:
-            element = await page.select(selector, timeout=timeout)
-            if element:
-                element = Element(element, css_selector=selector, config=config)
-                if skip_visibility_check or await element.is_visible():
-                    return element
-        return None
-    except (asyncio.TimeoutError, Exception):
-        return None
-
-
-async def page_batch_extract(
-    page: zd.Tab, queries: list[dict[str, object]]
-) -> dict[str, dict[str, object]] | None:
-    if len(queries) == 0:
-        return {}
-
-    payload = json.dumps(queries)
-    js_code = f"""
-    (() => {{
-        const queries = {payload};
-        const output = {{}};
-
-        function isVisible(element) {{
-            const style = window.getComputedStyle(element);
-            const rect = element.getBoundingClientRect();
-            return (
-                style.visibility !== "hidden" &&
-                style.display !== "none" &&
-                rect.width > 0 &&
-                rect.height > 0
-            );
-        }}
-
-        function findCss(selector, doc, iframeSelector) {{
-            try {{
-                if (iframeSelector) {{
-                    const iframe = doc.querySelector(iframeSelector);
-                    if (iframe) {{
-                        const childDoc = iframe.contentDocument || iframe.contentWindow.document;
-                        const directIframeMatch = childDoc.querySelector(selector);
-                        if (directIframeMatch) {{
-                            return directIframeMatch;
-                        }}
-                        return findCss(selector, childDoc, null);
-                    }}
-                }}
-            }} catch (error) {{
-                return null;
-            }}
-
-            try {{
-                const direct = doc.querySelector(selector);
-                if (direct) return direct;
-            }} catch (error) {{
-                return null;
-            }}
-
-            const iframes = doc.querySelectorAll("iframe");
-            for (const iframe of iframes) {{
-                try {{
-                    const childDoc = iframe.contentDocument || iframe.contentWindow.document;
-                    const nested = findCss(selector, childDoc, null);
-                    if (nested) return nested;
-                }} catch (error) {{}}
-            }}
-
-            return null;
-        }}
-
-        for (const query of queries) {{
-            const selector = query?.selector;
-            const iframeSelector = query?.iframe_selector;
-            const queryKey = query?.query_key;
-            if (typeof selector !== "string" || selector.length === 0) {{
-                continue;
-            }}
-            if (typeof queryKey !== "string" || queryKey.length === 0) {{
-                continue;
-            }}
-
-            let element = null;
-
-            if (selector.startsWith("//")) {{
-                try {{
-                    element = document.evaluate(
-                        selector,
-                        document,
-                        null,
-                        XPathResult.FIRST_ORDERED_NODE_TYPE,
-                        null
-                    ).singleNodeValue;
-                }} catch (error) {{
-                    element = null;
-                }}
-            }} else {{
-                const resolvedIframeSelector = typeof iframeSelector === "string" &&
-                    iframeSelector.length > 0
-                    ? iframeSelector
-                    : null;
-                element = findCss(selector, document, resolvedIframeSelector);
-            }}
-
-            if (!element || !isVisible(element)) {{
-                output[queryKey] = {{ found: false }};
-                continue;
-            }}
-
-            const tag = String(element.tagName || "").toLowerCase();
-            const item = {{ found: true, tag }};
-
-            if (query?.wants_html) {{
-                item.html = element.innerHTML || "";
-            }}
-            if (query?.wants_text) {{
-                item.text = element.innerText || "";
-            }}
-            if (query?.wants_value && ["input", "textarea", "select"].includes(tag)) {{
-                const value = element.value;
-                if (typeof value === "string") {{
-                    item.value = value;
-                }} else if (value == null) {{
-                    item.value = "";
-                }} else {{
-                    item.value = String(value);
-                }}
-            }}
-
-            output[queryKey] = item;
-        }}
-
-        return output;
-    }})()
-    """
-
-    try:
-        result = await page.evaluate(js_code)
-        if isinstance(result, dict):
-            return cast(dict[str, dict[str, object]], result)
-    except Exception as error:
-        logger.debug(f"Batch extract failed: {error}")
-    return None
 
 
 async def distill(
@@ -1048,13 +467,6 @@ async def autoclick(page: zd.Tab, distilled: str, expr: str):
                 logger.warning(f"Selector {selector} not found, can't click on it")
 
 
-async def get_url(page: zd.Tab) -> str | None:
-    current_url = await page.evaluate("window.location.href", await_promise=True)
-    if current_url:
-        return str(current_url)
-    return None
-
-
 async def run_distillation_loop(
     *,
     location: str | None = None,
@@ -1183,151 +595,3 @@ async def short_lived_mcp_tool(
                     ))
                     item["url"] = url
     return terminated, result
-
-
-async def page_batch_actions(page: zd.Tab, actions: list[dict[str, str]]) -> dict[str, bool] | None:
-    if len(actions) == 0:
-        return {}
-
-    payload = json.dumps(actions)
-    js_code = f"""
-    (async () => {{
-        const actions = {payload};
-        const output = {{}};
-
-        function findCss(selector, doc) {{
-            try {{
-                const direct = doc.querySelector(selector);
-                if (direct) return direct;
-            }} catch (error) {{
-                return null;
-            }}
-
-            const iframes = doc.querySelectorAll("iframe");
-            for (const iframe of iframes) {{
-                try {{
-                    const childDoc = iframe.contentDocument || iframe.contentWindow.document;
-                    const nested = findCss(selector, childDoc);
-                    if (nested) return nested;
-                }} catch (error) {{
-                    // Cross-origin iframe.
-                }}
-            }}
-            return null;
-        }}
-
-        for (const action of actions) {{
-            const key = action?.key;
-            const kind = action?.kind;
-            const selector = action?.selector;
-            if (typeof key !== "string" || key.length === 0) {{
-                continue;
-            }}
-            if (typeof selector !== "string" || selector.length === 0) {{
-                output[key] = false;
-                continue;
-            }}
-
-            let element = null;
-            if (selector.startsWith("//")) {{
-                try {{
-                    element = document.evaluate(
-                        selector,
-                        document,
-                        null,
-                        XPathResult.FIRST_ORDERED_NODE_TYPE,
-                        null
-                    ).singleNodeValue;
-                }} catch (error) {{
-                    element = null;
-                }}
-            }} else {{
-                element = findCss(selector, document);
-            }}
-
-            if (!element) {{
-                output[key] = false;
-                continue;
-            }}
-
-            try {{
-                if (kind === "click") {{
-                    element.click();
-                    output[key] = true;
-                }} else if (kind === "set_value") {{
-                    const value = typeof action?.value === "string" ? action.value : "";
-                    const requestedDelay = Number(action?.typing_delay_ms);
-                    const typingDelayMs = Number.isFinite(requestedDelay)
-                        ? Math.max(0, Math.min(250, requestedDelay))
-                        : 25;
-                    const sleep = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
-                    const nativeInputValueSetter = Object.getOwnPropertyDescriptor(
-                        window.HTMLInputElement.prototype,
-                        "value"
-                    )?.set;
-                    const nativeTextAreaValueSetter = Object.getOwnPropertyDescriptor(
-                        window.HTMLTextAreaElement.prototype,
-                        "value"
-                    )?.set;
-
-                    const setNativeValue = (el, nextValue) => {{
-                        if (el instanceof HTMLInputElement && nativeInputValueSetter) {{
-                            nativeInputValueSetter.call(el, nextValue);
-                        }} else if (el instanceof HTMLTextAreaElement && nativeTextAreaValueSetter) {{
-                            nativeTextAreaValueSetter.call(el, nextValue);
-                        }} else {{
-                            el.value = nextValue;
-                        }}
-                    }};
-
-                    element.focus();
-                    element.dispatchEvent(new KeyboardEvent("keydown", {{ key: "Tab", bubbles: true }}));
-
-                    setNativeValue(element, "");
-                    element.dispatchEvent(
-                        new InputEvent("input", {{
-                            bubbles: true,
-                            inputType: "deleteContentBackward",
-                            data: null
-                        }})
-                    );
-
-                    let currentValue = "";
-                    for (const char of value) {{
-                        element.dispatchEvent(new KeyboardEvent("keydown", {{ key: char, bubbles: true }}));
-                        element.dispatchEvent(new KeyboardEvent("keypress", {{ key: char, bubbles: true }}));
-                        currentValue += char;
-                        setNativeValue(element, currentValue);
-                        element.dispatchEvent(
-                            new InputEvent("input", {{
-                                bubbles: true,
-                                inputType: "insertText",
-                                data: char
-                            }})
-                        );
-                        element.dispatchEvent(new KeyboardEvent("keyup", {{ key: char, bubbles: true }}));
-                        if (typingDelayMs > 0) {{
-                            await sleep(typingDelayMs);
-                        }}
-                    }}
-                    element.dispatchEvent(new Event("change", {{ bubbles: true }}));
-                    output[key] = true;
-                }} else {{
-                    output[key] = false;
-                }}
-            }} catch (error) {{
-                output[key] = false;
-            }}
-        }}
-
-        return output;
-    }})()
-    """
-
-    try:
-        result = await page.evaluate(js_code, await_promise=True)
-        if isinstance(result, dict):
-            return {str(k): bool(v) for k, v in result.items()}  # pyright: ignore[reportUnknownArgumentType, reportUnknownVariableType]
-    except Exception as error:
-        logger.error(f"Batch actions failed: {error}")
-    return None
