+++ tests/distillation/test_distill.py
@@ -15,14 +15,17 @@ PROJECT_ROOT = Path(__file__).resolve().parents[2]
 
 ACME_HOSTNAME = "https://acme.fly.dev"
 
-from getgather.browser.chromefleet import create_remote_browser, terminate_remote_browser
+from getgather.browser import (
+    create_remote_browser,
+    get_new_page,
+    page_batch_extract,
+    terminate_remote_browser,
+)
 from getgather.config import FRIENDLY_CHARS, settings
 from getgather.zen_distill import (
     Pattern,
     distill,
-    get_new_page,
     load_distillation_patterns,
-    page_batch_extract,
     run_distillation_loop,
 )
